pyproject.toml
[tool.poetry]
name = "creditrack"
version = "0.1.0"
description = ""
authors = ["volodymyr-v-konoval <volodymyr.v.konoval@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.1"}
sqlalchemy = "^2.0.40"
asyncmy = "^0.2.10"
pydantic-settings = "^2.9.1"
pandas = "^2.2.3"
python-multipart = "^0.0.20"
cryptography = "^44.0.2"
alembic = "^1.15.2"
pymysql = "^1.1.1"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
